Update to version 0.0.12

+++ src/core/useFormPersistMultiplePages.ts
@@ -6,7 +6,7 @@ import type {
   UnpackNestedValue,
 } from 'react-hook-form';
 import { useEffect, useCallback } from 'react';
-import { isValidRecord, isValidRecords } from '../util';
+import { isValidRecord, isValidRecords, objectify, tryParseDate } from '../util';
 import { validate } from '../share';
 import type { NonEmptyString } from '../types';
 
@@ -47,19 +47,21 @@ export function useFormPersistMultiplePages<T extends FieldValues, U extends str
     const storage = getStorage();
     // Create an object from already storaged data
     const storaged = storage.getItem(ROOT_KEY);
-    const existed = storaged !== null ? JSON.parse(storaged) : {}; // eslint-disable-line
-    // Create an object for the additional data to be storaged
-    const removed = excludes.reduce((acc, key) => {
-      // FIXME: want to remove disable/ignore
-      // @ts-ignore
-      delete acc[key];
-      return acc;
-    }, inputted);
-    const key = dataKey || getPathname();
-    const added = { [key]: removed };
-    // Merge and storage them
-    const stringified = JSON.stringify({ ...existed, ...added });
-    storage.setItem(ROOT_KEY, stringified);
+    const existed: unknown = storaged === null ? {} : JSON.parse(storaged);
+    if (isValidRecord(existed)) {
+      // Create an object for the additional data to be storaged
+      const removed = excludes.reduce((acc, key) => {
+        // FIXME: want to remove ts-ignore
+        // @ts-ignore
+        delete acc[key];
+        return acc;
+      }, inputted);
+      const key = dataKey || getPathname();
+      const added = { [key]: removed };
+      // Merge and storage them
+      const stringified = JSON.stringify({ ...existed, ...added });
+      storage.setItem(ROOT_KEY, stringified);
+    }
   }, [ROOT_KEY, dataKey, excludes, inputted]);
 
   // Delete data in a storage
@@ -69,19 +71,21 @@ export function useFormPersistMultiplePages<T extends FieldValues, U extends str
   }, [ROOT_KEY]);
 
   // Retrieve all data from a storage and return them as an object
-  const getPersisted = useCallback(<V>() => {
+  const getPersisted = useCallback(<V extends FieldValues>() => {
     const storaged = getStorage().getItem(ROOT_KEY);
     if (storaged === null) {
       return null;
     }
     const parsed: unknown = JSON.parse(storaged);
     if (isValidRecord(parsed)) {
-      const values = Object.values(parsed);
-      if (isValidRecords(values)) {
-        const persisted = values.reduce((acc, obj) => {
-          return { ...acc, ...obj };
-        });
-        return persisted as V;
+      const extracted = Object.values(parsed);
+      if (isValidRecords(extracted)) {
+        const obj = objectify(extracted);
+        // Convert the value to Date if possible.
+        const datefied = Object.entries(obj).reduce<Record<string, unknown>>((acc, [k, v]) => {
+          return { ...acc, [k]: tryParseDate(v) };
+        }, {});
+        return datefied as V;
       }
     }
     return null;
